requirements.txt
smartapi-python
websocket-client
pyotp
pandas
python-dotenv
